Move sanitize functions to the core
